requirements.txt
matplotlib~=3.9.1.post1
pandas~=2.2.2
selenium~=4.22.0
requests~=2.32.3
bs4~=0.0.2
beautifulsoup4~=4.12.3
opencv-python~=4.10.0.84
numpy~=1.26.4
tensorflow~=2.17.0
seaborn~=0.13.2
pillow~=10.4.0
scipy~=1.14.1
scikit-learn~=1.5.1
albumentations~=1.4.14